# docs/swagger.yaml
definitions:
  dto.AddressCreateDTO:
    properties:
      address:
        type: string
      city:
        type: string
      is_primary:
        type: integer
      label:
        type: string
      postal_code:
        type: integer
      recipient_name:
        type: string
      recipient_phone:
        type: string
      user_id:
        type: integer
    required:
    - address
    - city
    - is_primary
    - label
    - postal_code
    - recipient_name
    - recipient_phone
    type: object
  dto.AddressUpdateDTO:
    properties:
      address:
        type: string
      city:
        type: string
      id:
        type: integer
      is_primary:
        type: integer
      label:
        type: string
      postal_code:
        type: integer
      recipient_name:
        type: string
      recipient_phone:
        type: string
      user_id:
        type: integer
    required:
    - address
    - city
    - is_primary
    - label
    - postal_code
    - recipient_name
    - recipient_phone
    type: object
...
  dto.CartCreateDTO:
    properties:
      product_id:
        type: integer
      qty:
        type: integer
      user_id:
        type: integer
    required:
    - product_id
    - qty
    type: object
  dto.CartUpdateDTO:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      qty:
        type: integer
      user_id:
        type: integer
    required:
    - product_id
    - qty
    type: object
...
  helper.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /Address:
    get:
      consumes:
      - application/json
      description: Get list Addresss
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
      summary: List Addresss
      tags:
      - Address
    post:
      consumes:
      - application/json
      description: Post Address
      parameters:
      - description: the body to create Address
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/dto.AddressCreateDTO'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Store Address
      tags:
      - Address
  /Address/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Address
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Delete Address
      tags:
      - Address
    get:
      consumes:
      - application/json
      description: Get detail Address
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Detail Address
      tags:
      - Address
    put:
      consumes:
      - application/json
      description: Update Address
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Address
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/dto.AddressUpdateDTO'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Update Address
      tags:
      - Address
...
  /cart:
    get:
      consumes:
      - application/json
      description: Get list carts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
      summary: List carts
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Post cart
      parameters:
      - description: the body to create cart
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/dto.CartCreateDTO'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Store cart
      tags:
      - Cart
  /cart/{id}:
    delete:
      consumes:
      - application/json
      description: Delete cart
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Delete cart
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Get detail cart
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Detail cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update cart
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update cart
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/dto.CartUpdateDTO'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Update cart
      tags:
      - Cart
...
swagger: "2.0"